break ACF class out
